libs: mission: extract confirm-button placement geometry
Move the shared screen-bounds and clamp-inside-viewport helpers that
position the survey polygon confirm button into a dedicated libs module,
where the upcoming geofence drawing action buttons can reuse them
without further inflating the planning view.
